+++ src/common/utils.f90
@@ -1,4 +1,4 @@
-! Copyright (c) 2019-2021, The Neko Authors
+! Copyright (c) 2019-2025, The Neko Authors
 ! All rights reserved.
 !
 ! Redistribution and use in source and binary forms, with or without
@@ -43,15 +43,94 @@ module utils
      module procedure neko_error_plain, neko_error_msg
   end interface neko_error
 
+  !! Interfaces for error and warning routines found in the
+  !! errors submodule.
+  interface
+     !> Reports an error and stops execution.
+     !! @param[optional] error_code The error code to report.
+     module subroutine neko_error_plain(error_code)
+       integer, optional, intent(in) :: error_code
+     end subroutine neko_error_plain
+
+     !> Reports an error and stops execution.
+     !! @param error_msg The error message to report.
+     module subroutine neko_error_msg(error_msg)
+       character(len=*), intent(in) :: error_msg
+     end subroutine neko_error_msg
+     !> Reports a type selection error and stops execution.
+     !! @param base_type The base type for which the selection was attempted.
+     !! @param wrong_type The type that was attempted to be selected.
+     !! @param known_types An array of known valid types.
+     module subroutine neko_type_error(base_type, wrong_type, known_types)
+       character(len=*), intent(in) :: base_type
+       character(len=*), intent(in) :: wrong_type
+       character(len=*), intent(in) :: known_types(:)
+     end subroutine neko_type_error
+
+     !> Reports a type registration error and stops execution.
+     !! @param base_type The base type for which the registration was attempted.
+     !! @param wrong_type The type that was attempted to be registered.
+     !! @param known Whether the conflicting type is known (standard) or custom.
+     module subroutine neko_type_registration_error(base_type, wrong_type, &
+          known)
+       character(len=*), intent(in) :: base_type
+       character(len=*),intent(in) :: wrong_type
+       logical, intent(in) :: known
+     end subroutine neko_type_registration_error
+
+     !> Reports a warning to standard output
+     !! @param warning_msg The warning message to report.
+     module subroutine neko_warning(warning_msg)
+       character(len=*) :: warning_msg
+     end subroutine neko_warning
+  end interface
+
+  abstract interface
+     !> Interface for the throw procedure. Follows pFunit conventions.
+     subroutine throw_intf(filename, line_number, message)
+       character(len=*), intent(in) :: filename
+       integer, intent(in) :: line_number
+       character(len=*), optional, intent(in) :: message
+     end subroutine throw_intf
+  end interface
+
+  !> Pointer to the throw procedure.
+  !! @details Ordinarily only raises an error stop. During testing,
+  !! pFUnit will hijack this procedure to raise an excpeption that
+  !! can be caught by the testing framework.
+  procedure(throw_intf), pointer :: throw_error => null()
+  !> Same as above, but for warnings. Does nothing by default
+  procedure(throw_intf), pointer :: throw_warning => null()
+
   public :: neko_error, neko_warning, nonlinear_index, filename_chsuffix, &
        filename_path, filename_name, filename_suffix, &
        filename_suffix_pos, filename_tslash_pos, &
        linear_index, split_string, NEKO_FNAME_LEN, index_is_on_facet, &
        concat_string_array, extract_fld_file_index, neko_type_error, &
-       neko_type_registration_error
+       neko_type_registration_error, throw_error, throw_warning, throw_intf, &
+       default_throw_error, default_throw_warning
 
 
 contains
+
+  !> Default throw method for warnings. Does nothing.
+  subroutine default_throw_warning(filename, line_number, message)
+    character(len=*), intent(in) :: filename
+    integer, intent(in) :: line_number
+    character(len=*), optional, intent(in) :: message
+  end subroutine default_throw_warning
+
+  !> Default throw method that stops execution.
+  !! @details pFUnit will highjack this method during testing to catch
+  !! exceptions.
+  subroutine default_throw_error(filename, line_number, message)
+    character(len=*), intent(in) :: filename
+    integer, intent(in) :: line_number
+    character(len=*), optional, intent(in) :: message
+
+    error stop
+  end subroutine default_throw_error
+
 
   !> Find position (in the string) of a filename's suffix
   pure function filename_suffix_pos(fname) result(suffix_pos)
@@ -279,73 +358,6 @@ contains
     end select
 
   end function index_is_on_facet
-
-  !> Reports an error and stops execution
-  !! @param[optional] error_code The error code to report.
-  subroutine neko_error_plain(error_code)
-    integer, optional :: error_code
-
-    if (present(error_code)) then
-       write(error_unit, *) '*** ERROR ***', error_code
-       error stop
-    else
-       write(error_unit, *) '*** ERROR ***'
-       error stop
-    end if
-
-  end subroutine neko_error_plain
-
-  !> Reports an error and stops execution
-  !! @param error_msg The error message to report.
-  subroutine neko_error_msg(error_msg)
-    character(len=*) :: error_msg
-    write(error_unit, *) '*** ERROR: ', error_msg, ' ***'
-    error stop
-  end subroutine neko_error_msg
-
-  !> Reports an error allocating a type for a particular base pointer class.
-  !! @details Should be used in factories.
-  !! @param base_type The base type of the object, which the factory tried to
-  !! construct.
-  !! @param wrong_type The type that was attempted to construct.
-  !! @param known_types A list of the types that are known.
-  subroutine neko_type_error(base_type, wrong_type, known_types)
-    character(len=*), intent(in) :: base_type
-    character(len=*), intent(in) :: wrong_type
-    character(len=*), intent(in) :: known_types(:)
-    integer :: i
-
-    write(error_unit, *) '*** ERROR WHEN SELECTING TYPE ***'
-    write(error_unit, *) 'Type ', wrong_type, ' does not exist for ', base_type
-    write(error_unit, *) 'Valid types are:'
-    do i = 1, size(known_types)
-       write(error_unit, *) "    ", known_types(i)
-    end do
-    error stop
-  end subroutine neko_type_error
-
-  subroutine neko_type_registration_error(base_type, wrong_type, known)
-    character(len=*), intent(in) :: base_type
-    character(len=*),intent(in) :: wrong_type
-    logical, intent(in) :: known
-
-    write(error_unit, *) '*** ERROR WHEN REGISTERING TYPE ***'
-    write(error_unit, *) 'Type name ', wrong_type, &
-         ' conflicts with and already existing ', base_type, " type"
-    if (known) then
-       write(error_unit, *) 'Please rename your custom type.'
-    else
-       write(error_unit, *) 'The already existing type is also custom.' // &
-            ' Make all custom type names unique!'
-    end if
-    error stop
-  end subroutine neko_type_registration_error
-
-  !> Reports a warning to standard output
-  subroutine neko_warning(warning_msg)
-    character(len=*) :: warning_msg
-    write(output_unit, *) '*** WARNING: ', warning_msg, ' ***'
-  end subroutine neko_warning
 
   !> Concatenate an array of strings into one string with array items
   !! separated by spaces.